refactor: deduplicate xUnit2 TestKit support
Share the Akka.Hosting.TestKit implementation between the xUnit 3 and xUnit 2 packages and keep the xUnit 2 test project focused on compatibility coverage. This preserves the migration work while cutting the duplicated maintenance surface introduced by #728.
