--- src/Akka.Hosting.TestKit.Xunit2.Tests/TestActorRefTests/TActorBase.cs
@@ -1,30 +1,0 @@
-﻿//-----------------------------------------------------------------------
-// <copyright file="TActorBase.cs" company="Akka.NET Project">
-//     Copyright (C) 2009-2021 Lightbend Inc. <http://www.lightbend.com>
-//     Copyright (C) 2013-2021 .NET Foundation <https://github.com/akkadotnet/akka.net>
-// </copyright>
-//-----------------------------------------------------------------------
-
-using System.Threading;
-using Akka.Actor;
-
-namespace Akka.Hosting.TestKit.Tests.TestActorRefTests;
-
-// ReSharper disable once InconsistentNaming
-public abstract class TActorBase : ActorBase
-{
-    protected sealed override bool Receive(object message)
-    {
-        var currentThread = Thread.CurrentThread;
-        if(currentThread != TestActorRefSpec.Thread)
-            TestActorRefSpec.OtherThread = currentThread;
-        return ReceiveMessage(message);
-    }
-
-    protected abstract bool ReceiveMessage(object message);
-
-    protected ActorSystem System
-    {
-        get { return ((LocalActorRef)Self).Cell.System; }
-    }
-}
