--- src/Akka.Hosting.TestKit.Xunit2.Tests/TestActorRefTests/WrappedTerminated.cs
@@ -1,22 +1,0 @@
-﻿//-----------------------------------------------------------------------
-// <copyright file="WrappedTerminated.cs" company="Akka.NET Project">
-//     Copyright (C) 2009-2021 Lightbend Inc. <http://www.lightbend.com>
-//     Copyright (C) 2013-2021 .NET Foundation <https://github.com/akkadotnet/akka.net>
-// </copyright>
-//-----------------------------------------------------------------------
-
-using Akka.Actor;
-
-namespace Akka.Hosting.TestKit.Tests.TestActorRefTests;
-
-public class WrappedTerminated
-{
-    private readonly Terminated _terminated;
-
-    public WrappedTerminated(Terminated terminated)
-    {
-        _terminated = terminated;
-    }
-
-    public Terminated Terminated { get { return _terminated; } }
-}
