+++ src/Akka.Hosting.TestKit.Xunit2/GlobalUsings.cs
@@ -1,0 +1,2 @@
+global using XunitAssertions = Akka.TestKit.Xunit2.XunitAssertions;
+global using XunitTestOutputHelper = Xunit.Abstractions.ITestOutputHelper;
