+++ src/Akka.Hosting.TestKit/GlobalUsings.cs
@@ -1,0 +1,2 @@
+global using XunitAssertions = Akka.TestKit.Xunit.XunitAssertions;
+global using XunitTestOutputHelper = Xunit.ITestOutputHelper;
